% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/france_deps.R
\docType{data}
\name{france_deps}
\alias{france_deps}
\title{French Departments outline collected from data.gouv.fr}
\format{
An sf data.frame with 96 observation and 5 variables:
\describe{
\item{name}{Name identifying department}
\item{name_prefecture}{Name identifying préfecture of department}
\item{x}{Long of department centroid}
\item{y}{Lat of department centroid}
\item{geometry}{MULTIPOLYGON representing department outline}
}
}
\source{
\url{https://www.data.gouv.fr/en/datasets/carte-des-departements-2-1/}
}
\usage{
france_deps
}
\description{
Outline of French departments collected from data.gouv.fr, simplified and in sf geometry.
}
\keyword{datasets}
