% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_locations.R
\name{grid_locations}
\alias{grid_locations}
\title{grid_locations}
\usage{
grid_locations(n_row, n_col, spacers = list())
}
\arguments{
\item{n_row}{maximum number of rows defining grid. There must be at least as many grid cells as there are points to allocate.}

\item{n_col}{maximum number of columns defining grid. There must be at least as many grid cells as there are points to allocate.}

\item{spacers}{Optional list of grid cell locations defining grid location of fixed spacers which cannot be allocated points. Coordinates are in (row, column) order with origin top-left. Default is an empty array.}
}
\value{
A tibble of grid coordinates in the same order as the original array of input points.
}
\description{
From https://observablehq.com/@jwolondon/gridmap-allocation.
Creates a list of grid locations of a stated size, with spacer positions
removed.
}
\examples{
grid_locations(n_row = 3, n_col = 3, spacers = list())
}
