% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/london_boroughs.R
\docType{data}
\name{london_boroughs}
\alias{london_boroughs}
\title{London borough outline collected from OS OpenGeography}
\format{
An sf data.frame with 33 observation and 5 variables:
\describe{
\item{area_code}{Code identifying local authority}
\item{area_name}{Name identifying local authority}
\item{easting}{Easting of local authority centroid}
\item{northing}{Northing of local authority centroid}
\item{geometry}{MULTIPOLYGON representing local authority outline}
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
london_boroughs
}
\description{
Outline of London Boroughs collected from OS OpenGeography, simplified and in sf geometry.
}
\keyword{datasets}
