% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_lp.R
\name{solve_lp}
\alias{solve_lp}
\title{solve_lp}
\usage{
solve_lp(pts, grd, compactness)
}
\arguments{
\item{pts}{tibble of geographic points (x,y) to be transformed for allocation to grid.}

\item{grd}{tibble defining grid positions.}

\item{compactness}{Optional parameter between 0 and 1 where 0 allocates towards edges, 0.5 preserves scaled geographic location and 1 allocates towards centre of grid. Default is 1 (compact cluster).}
}
\value{
A tibble containing solution with row-ids of pts and grd tibbles on which model constructed.
}
\description{
From https://observablehq.com/@jwolondon/gridmap-allocation.
Returns an LP solution allocating 2D points to a grid.
}
\examples{

library(sf)
library(dplyr)
pts <- london_boroughs |> st_drop_geometry() |> select(area_name, x = easting, y = northing)
grd <- grid_locations(n_row = 8, n_col = 8, spacers = list())
grd_pos <- solve_lp(points_normalised(pts, grd, 0.6), grd, 0.6)
grd_pos

}
\author{
Roger Beecham
}
