% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_lp.R
\name{solve_lp}
\alias{solve_lp}
\title{solve_lp}
\usage{
solve_lp(pts, grd, compactness)
}
\arguments{
\item{pts}{tibble of geographic points (x,y) to be transformed for allocation to grid.}

\item{grd}{tibble defining grid positions.}

\item{compactness}{Optional parameter between 0 and 1 where 0 allocates towards edges, 0.5 preserves scaled geographic location and 1 allocates towards centre of grid. Default is 1 (compact cluster).}
}
\value{
A tibble containing solution with row-ids of pts and grd tibbles on which model constructed.
}
\description{
From https://observablehq.com/@jwolondon/gridmap-allocation.
Returns an LP solution allocating 2D points to a grid.
}
\examples{
library(tibble)
pts <- tribble(
  ~x, ~y,
  2, 4,
  1, 5,
  2, 1,
  3, 3,
  3, 4
)
grd <- grid_locations(n_row = 3, n_col = 3, spacers = list())
solve_lp(pts, grd, 0.6)
}
